fix(#1909): PR #2016 — clear structural file-size ratchets by extracting cohesive blocks into siblings
The write-hot-paths additions pushed five remnic-core files past their
grandfathered ceilings (storage.ts 7061->7397, config.ts 4560->4601,
orchestrator.ts 3979->4004, orchestration/extraction-persist.ts
2998->3019 crossing the 3000 oversize threshold, types.ts 3870->3878).

Apply the ratchet's prescribed remedy — extract cohesive blocks into new
sibling modules and re-import — rather than raising baselines:

- storage.ts: entity timeline/structured-section pure helpers ->
  storage/entity-timeline.ts (moved exports re-exported for a stable API).
- config.ts: QMD config parsers -> config-qmd.ts.
- types.ts: continuity incident/loop types -> types-continuity.ts (re-exported).
- orchestrator.ts: semanticDedupScopeFor + qmdCollectionNamespaceFromPrefix ->
  orchestration/orchestrator-namespace-scope.ts as free functions; the scope
  helper takes a structural { dir } param, dropping the direct storage.js import.
- extraction-persist.ts: auto-promotion decision helpers ->
  orchestration/extraction-persist-promotion.ts as free functions.

Verbatim moves; behavior unchanged. No ratchet-baseline edits.
tsc --noEmit clean; focused tests green (443 pass).

Now under every ceiling: storage 6953, config 4521, orchestrator 3953,
extraction-persist 2967, types 3834; storage.js direct imports back to 42.
