+++ packages/remnic-core/src/types.ts
@@ -3630,63 +3630,19 @@ export interface CompressionGuidelineOptimizerState {
   ruleUpdates?: CompressionGuidelineOptimizerRuleUpdate[];
 }
 
-export type ContinuityIncidentState = "open" | "closed";
-
-export interface ContinuityIncidentRecord {
-  id: string;
-  state: ContinuityIncidentState;
-  openedAt: string;
-  updatedAt: string;
-  triggerWindow?: string;
-  symptom: string;
-  suspectedCause?: string;
-  fixApplied?: string;
-  verificationResult?: string;
-  preventiveRule?: string;
-  closedAt?: string;
-  filePath?: string;
-}
-
-export interface ContinuityIncidentOpenInput {
-  triggerWindow?: string;
-  symptom: string;
-  suspectedCause?: string;
-}
-
-export interface ContinuityIncidentCloseInput {
-  fixApplied: string;
-  verificationResult: string;
-  preventiveRule?: string;
-}
-
-export type ContinuityLoopCadence = "daily" | "weekly" | "monthly" | "quarterly";
-export type ContinuityLoopStatus = "active" | "paused" | "retired";
-
-export interface ContinuityImprovementLoop {
-  id: string;
-  cadence: ContinuityLoopCadence;
-  purpose: string;
-  status: ContinuityLoopStatus;
-  killCondition: string;
-  lastReviewed: string;
-  notes?: string;
-}
-
-export interface ContinuityLoopUpsertInput {
-  id: string;
-  cadence: ContinuityLoopCadence;
-  purpose: string;
-  status: ContinuityLoopStatus;
-  killCondition: string;
-  lastReviewed?: string;
-  notes?: string;
-}
-
-export interface ContinuityLoopReviewInput {
-  status?: ContinuityLoopStatus;
-  notes?: string;
-  reviewedAt?: string;
-}
+// Continuity incident + improvement-loop types live in a sibling module to keep
+// types.ts under its size ceiling; re-exported so `./types.js` imports resolve.
+export type {
+  ContinuityIncidentState,
+  ContinuityIncidentRecord,
+  ContinuityIncidentOpenInput,
+  ContinuityIncidentCloseInput,
+  ContinuityLoopCadence,
+  ContinuityLoopStatus,
+  ContinuityImprovementLoop,
+  ContinuityLoopUpsertInput,
+  ContinuityLoopReviewInput,
+} from "./types-continuity.js";
 
 /** Entry in the access tracking buffer (batched updates) */
 export interface AccessTrackingEntry {
